Cleanup base plugin supposed to be applied by all projects
