+++ build-tools-internal/src/main/groovy/elasticsearch.base.gradle
@@ -1,0 +1,14 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0 and the Server Side Public License, v 1; you may not use this file except
+ * in compliance with, at your election, the Elastic License 2.0 or the Server
+ * Side Public License, v 1.
+ */
+
+import org.elasticsearch.gradle.VersionProperties
+
+project.setDescription("Elasticsearch subproject " + project.getPath());
+// common maven publishing configuration
+project.setGroup("org.elasticsearch");
+project.setVersion(VersionProperties.getElasticsearch())
